load indices and buffers using cgltf unpack functions instead

+++ include/models/model.h
@@ -12,7 +12,7 @@ typedef struct {
   float* normals;
   float* tangents;
   float* tex_coords;
-  unsigned short* indices;
+  unsigned int* indices;
 } Mesh;
 
 typedef struct {
